refactor: split NewMessageEventHandler
