+++ logic/src/commonTest/kotlin/com/wire/kalium/logic/framework/TestEvent.kt
@@ -1,5 +1,6 @@
 package com.wire.kalium.logic.framework
 
+import com.wire.kalium.cryptography.utils.EncryptedData
 import com.wire.kalium.logic.data.conversation.ClientId
 import com.wire.kalium.logic.data.conversation.Conversation.Member
 import com.wire.kalium.logic.data.conversation.MutedConversationStatus
@@ -7,6 +8,7 @@ import com.wire.kalium.logic.data.event.Event
 import com.wire.kalium.logic.data.user.Connection
 import com.wire.kalium.logic.data.user.ConnectionState
 import com.wire.kalium.logic.data.user.UserId
+import kotlinx.datetime.Instant
 
 object TestEvent {
 
@@ -97,5 +99,29 @@ object TestEvent {
         teamId = "teamId",
         memberId = "memberId",
         permissionCode = permissionCode
+    )
+
+    fun newMessageEvent(
+        encryptedContent: String,
+        senderUserId: UserId = TestUser.USER_ID,
+        encryptedExternalContent: EncryptedData? = null
+    ) = Event.Conversation.NewMessage(
+        "eventId",
+        TestConversation.ID,
+        senderUserId,
+        TestClient.CLIENT_ID,
+        "time",
+        encryptedContent,
+        encryptedExternalContent
+    )
+
+    fun newMLSMessageEvent(
+        timestamp: Instant
+    ) = Event.Conversation.NewMLSMessage(
+        "eventId",
+        TestConversation.ID,
+        TestUser.USER_ID,
+        timestamp.toString(),
+        "content"
     )
 }
