+++ logic/src/commonTest/kotlin/com/wire/kalium/logic/sync/receiver/conversation/message/ProteusMessageUnpackerTest.kt
@@ -1,0 +1,142 @@
+package com.wire.kalium.logic.sync.receiver.conversation.message
+
+import com.wire.kalium.cryptography.CryptoClientId
+import com.wire.kalium.cryptography.CryptoSessionId
+import com.wire.kalium.cryptography.CryptoUserID
+import com.wire.kalium.cryptography.ProteusClient
+import com.wire.kalium.cryptography.utils.PlainData
+import com.wire.kalium.cryptography.utils.encryptDataWithAES256
+import com.wire.kalium.cryptography.utils.generateRandomAES256Key
+import com.wire.kalium.logic.data.message.MessageContent
+import com.wire.kalium.logic.data.message.PlainMessageBlob
+import com.wire.kalium.logic.data.message.ProtoContent
+import com.wire.kalium.logic.data.message.ProtoContentMapper
+import com.wire.kalium.logic.feature.ProteusClientProvider
+import com.wire.kalium.logic.framework.TestEvent
+import com.wire.kalium.logic.functional.Either
+import com.wire.kalium.logic.util.Base64
+import com.wire.kalium.logic.util.shouldSucceed
+import com.wire.kalium.protobuf.encodeToByteArray
+import com.wire.kalium.protobuf.messages.GenericMessage
+import com.wire.kalium.protobuf.messages.Text
+import io.ktor.utils.io.core.toByteArray
+import io.mockative.Mock
+import io.mockative.any
+import io.mockative.classOf
+import io.mockative.eq
+import io.mockative.given
+import io.mockative.matchers.Matcher
+import io.mockative.matching
+import io.mockative.mock
+import io.mockative.once
+import io.mockative.verify
+import kotlinx.coroutines.test.runTest
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertIs
+
+class ProteusMessageUnpackerTest {
+
+    @Test
+    fun givenNewMessageEvent_whenUnpacking_shouldAskProteusClientForDecryption() = runTest {
+        val (arrangement, proteusUnpacker) = Arrangement()
+            .withProteusClientDecryptingByteArray(decryptedData = byteArrayOf())
+            .withProtoContentMapperReturning(any(), ProtoContent.Readable("uuid", MessageContent.Unknown()))
+            .arrange()
+
+        val encodedEncryptedContent = Base64.encodeToBase64("Hello".encodeToByteArray())
+        val messageEvent = TestEvent.newMessageEvent(encodedEncryptedContent.decodeToString())
+        proteusUnpacker.unpackProteusMessage(messageEvent)
+
+        val cryptoSessionId = CryptoSessionId(
+            CryptoUserID(messageEvent.senderUserId.value, messageEvent.senderUserId.domain),
+            CryptoClientId(messageEvent.senderClientId.value)
+        )
+
+        val decodedByteArray = Base64.decodeFromBase64(messageEvent.content.toByteArray())
+        verify(arrangement.proteusClient)
+            .suspendFunction(arrangement.proteusClient::decrypt)
+            .with(matching { it.contentEquals(decodedByteArray) }, eq(cryptoSessionId))
+            .wasInvoked(exactly = once)
+    }
+
+    @Test
+    fun givenNewMessageEventWithExternalContent_whenUnpacking_shouldReturnDecryptedExternalMessage() = runTest {
+        val aesKey = generateRandomAES256Key()
+        val messageUid = "uuid"
+        val externalInstructions = ProtoContent.ExternalMessageInstructions(
+            messageUid,
+            aesKey.data,
+            sha256 = null,
+            encryptionAlgorithm = null
+        )
+        val plainTextContent = "Hello!"
+
+        val protobufExternalContent = GenericMessage(
+            content = GenericMessage.Content.Text(Text(plainTextContent)),
+            messageId = messageUid
+        )
+        val encryptedProtobufExternalContent = encryptDataWithAES256(PlainData(protobufExternalContent.encodeToByteArray()), aesKey)
+        val decryptedExternalContent = MessageContent.Text(plainTextContent)
+        val emptyArray = byteArrayOf()
+
+        val (_, proteusUnpacker) = Arrangement()
+            .withProteusClientDecryptingByteArray(decryptedData = emptyArray)
+            .withProtoContentMapperReturning(matching { it.data.contentEquals(emptyArray) }, externalInstructions)
+            .withProtoContentMapperReturning(
+                matching { it.data.contentEquals(protobufExternalContent.encodeToByteArray()) },
+                ProtoContent.Readable(messageUid, decryptedExternalContent)
+            ).arrange()
+
+        val messageEvent = TestEvent.newMessageEvent(
+            Base64.encodeToBase64("anything".encodeToByteArray()).decodeToString(),
+            encryptedExternalContent = encryptedProtobufExternalContent
+        )
+
+        val result = proteusUnpacker.unpackProteusMessage(messageEvent)
+
+        result.shouldSucceed {
+            assertIs<MessageUnpackResult.ApplicationMessage>(it)
+            val content = it.content
+            assertIs<ProtoContent.Readable>(content)
+            assertEquals(decryptedExternalContent, content.messageContent)
+        }
+    }
+
+    private class Arrangement {
+        @Mock
+        val proteusClient = mock(classOf<ProteusClient>())
+
+        @Mock
+        val proteusClientProvider = mock(classOf<ProteusClientProvider>())
+
+        @Mock
+        val protoContentMapper = mock(classOf<ProtoContentMapper>())
+
+        init {
+            given(proteusClientProvider)
+                .suspendFunction(proteusClientProvider::getOrError)
+                .whenInvoked()
+                .thenReturn(Either.Right(proteusClient))
+        }
+
+        fun withProteusClientDecryptingByteArray(decryptedData: ByteArray) = apply {
+            given(proteusClient)
+                .suspendFunction(proteusClient::decrypt)
+                .whenInvokedWith(any(), any())
+                .thenReturn(decryptedData)
+        }
+
+        fun withProtoContentMapperReturning(plainBlobMatcher: Matcher<PlainMessageBlob>, protoContent: ProtoContent) = apply {
+            given(protoContentMapper)
+                .function(protoContentMapper::decodeFromProtobuf)
+                .whenInvokedWith(plainBlobMatcher)
+                .thenReturn(protoContent)
+        }
+
+        fun arrange() = this to ProteusMessageUnpackerImpl(
+            proteusClientProvider, protoContentMapper
+        )
+    }
+
+}
